backup_v1/src/foip/operations_reconciliation_core.py
"""Operations reconciliation core inside foip package."""

import asyncio
import uuid
from datetime import datetime, timezone
from decimal import Decimal
from typing import Any, Dict, List, Optional

import attr

from foip.shared_metadata_schema import (
    EntityID,
    EntityType,
    BankTransaction,
    TimePoint,
    Company,
)
from foip.shared_metadata_registry import EntityRegistry

import structlog
log = structlog.get_logger(__name__)


class BankReconciliationEngine:
    def __init__(self, entity_registry: EntityRegistry, bank_connector: Any, transaction_matcher: Any):
        self.entity_registry = entity_registry
        self.bank_connector = bank_connector
        self.matcher = transaction_matcher
        self.initialized = False
        self.matching_rules = [
            {"type": "exact", "fields": ["amount", "date", "counterparty"]},
            {"type": "fuzzy", "fields": ["amount", "counterparty"], "tolerance": 0.01},
            {"type": "date_range", "fields": ["amount"], "days": 3},
        ]
        log.info("Bank Reconciliation Engine initialized")
