backup_v1/requirements.txt

# Core libraries
attrs>=23.1.0
msgspec>=0.18.0
structlog>=23.0.0
pydantic>=2.5.0

# Async I/O
aiohttp>=3.9.0
aiofiles>=23.2.0
anyio>=5.0.0
watchdog>=2.1.0
schedule>=1.3.0
tenacity>=8.2.0

# Caching & storage
cachetools>=5.3.0
diskcache>=5.6.0
redis>=5.0.0
asyncpg>=0.29.0

# Database / Migrations
sqlalchemy>=2.0.0
alembic>=1.13.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Monitoring & Logging
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.0
pytest-cov>=4.1.0
pluggy>=1.0.0

# Development / Tooling
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0

# Note: some adapters (e.g., foip_msgspec_adapter) are in-repo; they do not need
# separate PyPI packages. Lock exact versions in CI as needed.
